Fix user should be able to reset their own password
Yet to be test in our reset tool is using the app user to change password or
the current user.

if it use the app user, we should ensure the current user needs to provide
the current password to avoid any user change password user from ip
allow to access to write ldap server

if it use the current user, we needs to make sure the write server is
not accessible from users.